push assert scripts
